+++ cofi-wearos/src/main/java/com/omelan/cofi/wearos/presentation/pages/RecipeDetails.kt
@@ -1,0 +1,50 @@
+package com.omelan.cofi.wearos.presentation.pages
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.wear.compose.material.Button
+import androidx.wear.compose.material.CircularProgressIndicator
+import androidx.wear.compose.material.Icon
+import androidx.wear.compose.material.Text
+import com.omelan.cofi.share.Recipe
+import com.omelan.cofi.share.RecipeViewModel
+import com.omelan.cofi.share.StepsViewModel
+import com.omelan.cofi.wearos.R
+
+@Composable
+fun RecipeDetails(recipeId: Int) {
+    val recipeViewModel: RecipeViewModel = viewModel()
+    val stepsViewModel: StepsViewModel = viewModel()
+    val recipe by recipeViewModel.getRecipe(recipeId).observeAsState(initial = Recipe(name = ""))
+    val steps by stepsViewModel.getAllStepsForRecipe(recipeId).observeAsState(emptyList())
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center,
+    ) {
+        CircularProgressIndicator(
+            modifier = Modifier.fillMaxSize(),
+            progress = 0.5f,
+        )
+        Column(
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally,
+        ) {
+            Text(text = recipe.name)
+            Button(onClick = { /*TODO*/ }) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_gavel),
+                    contentDescription = null,
+                )
+            }
+        }
+    }
+}
